Fix the INSTALL documentation
I accidentally muddled the compile/install documentation with a previous
commit. I didn't realize there was a separate INSTALL-FROM-GIT file and
mistakenly updated INSTALL. To remedy this I updated README.md to
include the compile from git instructions. README is the instructions
for the tarball source. Also, I updated the root Makefile so that README.md
will not end up in the `make dist` tarball.