Fix for __nesc_keyword_ function undeclared
Added a check to semantics.c (function lookup_id()) to see if the
function name starts with __nesc_keyword_.  If so, nesC now checks to
see if the suffix function name is defined before throwing an error.

Also moved the #define for __nesc_keyword_ to a header file that both
relevant files include.

Fixes issue #4